move global context wrapper directory.
