+++ frontend/src/App.tsx
@@ -1,9 +1,9 @@
 // import React, { useState } from 'react'
 import { AppRouter } from '@/AppRouter'
-import { GlobalContextWrapper } from '@/components/_global/GlobalContextWrapper'
 import { GlobalFooter } from '@/components/_global/GlobalFooter'
 import { GlobalHeader } from '@/components/_global/GlobalHeader'
 import { GlobalBackToPageTopButton } from '@/components/_global/GlobalBackToPageTopButton'
+import { GlobalContextWrapper } from '@/components/_global/context/GlobalContextWrapper'
 import { AuthAppProviderContainer } from '@/components/container/AuthAppProviderContainer'
 import { GlobalLoadingProviderContainer } from '@/components/container/GlobalLoadingProviderContainer'
 import { ToastProviderContainer } from '@/components/container/ToastProviderContainer'
